app/requirements_app.txt
click==7.1.2
yarl==1.6.3
peewee==3.14.4
pandas==1.2.4
instaloader==4.7.1
PySimpleGUI==4.41.2
PyYAML==5.4.1
psycopg2-binary
requests
sentry_sdk